update README and refactor src
