+++ packages/express/src/node/index.ts
@@ -1,4 +1,4 @@
 export * from "./jwtSerDe";
 export * from "./redisPubSub";
 export * from "./server";
-export * from "./wsLiveViewAdaptor";
+export * from "./wsAdaptor";
